feat: nous camps en alta ciclista

+++ client/src/views/Home.vue
@@ -7,16 +7,16 @@
         <th scope="col">Birth Date</th>
         <th scope="col">Height</th>
         <th scope="col">weight</th>
-        <th scope="col">Delete?</th>
+        <th scope="col"></th>
       </tr>
     </thead>
     <tbody id="afegirAqui">
       <tr v-for="cyclist in cyclists" :key="cyclist.id">
         <td>{{ cyclist._id.substr(0, 5) }}</td>
         <td>{{ cyclist.name }}</td>
-        <td>{{ String(cyclist.birth_date).substr(0, 9) }}</td>
-        <td>{{ cyclist.height }}</td>
-        <td>{{ cyclist.weight }}</td>
+        <td>{{ formataData(String(cyclist.birth_date).substr(0, 10)) }}</td>
+        <td>{{ formataHeight(String(cyclist.height)) }} m.</td>
+        <td>{{ cyclist.weight }} Kg.</td>
         <td>
           <button @click="removeCyclist(cyclist._id)">Delete</button>
         </td>
@@ -62,6 +62,9 @@ export default {
     */
     removeCyclist(id) {
       // Delete in Collection
+      if (!window.confirm("Estàs segur que vols eliminar aquest ciclista?"))
+        return
+
       fetch(`${this.api_url}/cyclists/delete/${id}`, {
         method: 'DELETE'
       })
@@ -80,8 +83,22 @@ export default {
         return id == cyclist._id
       })
       this.cyclists.splice(index, 1)
-    }
+    },
+
+    formataData(data_in) {
+      const mesos = ["Gen", "Feb", "Mar", "Abr", "Mai", "Jun", "Jul", "Ago", "Sep", "Oct", "Nov", "Des"]
+      let dia = data_in.substring(8, 10)
+      let mes = mesos[data_in.substring(5, 7) - 1]
+      let any = data_in.substring(0, 4)
+
+      return `${dia}-${mes}-${any}`
+    },
+
+    formataHeight(centimetres) { return `${centimetres[0]}.${centimetres.substring(1, 3)}` }
+
   },
+
+
 
 
   // On mounted
